Make it possible to supply custom dispatcher implementations
