+++ piwik-sdk/src/main/java/org/piwik/sdk/Tracker.java
@@ -469,7 +469,7 @@ public class Tracker {
         } else {
             try {
                 // Another thread might be creating a sessions first transmission.
-                mSessionStartLatch.await(getDispatchTimeout(), TimeUnit.MILLISECONDS);
+                mSessionStartLatch.await(mDispatcher.getConnectionTimeOut(), TimeUnit.MILLISECONDS);
             } catch (InterruptedException e) { Timber.tag(TAG).e(e, null); }
         }
