+++ piwik-sdk/src/main/java/org/piwik/sdk/dispatcher/DefaultDispatcherFactory.java
@@ -1,0 +1,14 @@
+package org.piwik.sdk.dispatcher;
+
+import org.piwik.sdk.Tracker;
+import org.piwik.sdk.tools.Connectivity;
+
+public class DefaultDispatcherFactory implements DispatcherFactory {
+    public Dispatcher build(Tracker tracker) {
+        return new DefaultDispatcher(
+                new EventCache(new EventDiskCache(tracker)),
+                new Connectivity(tracker.getPiwik().getContext()),
+                new PacketFactory(tracker.getAPIUrl())
+        );
+    }
+}
