+++ piwik-sdk/src/main/java/org/piwik/sdk/dispatcher/DispatcherFactory.java
@@ -1,15 +1,7 @@
 package org.piwik.sdk.dispatcher;
 
 import org.piwik.sdk.Tracker;
-import org.piwik.sdk.tools.Connectivity;
 
-public class DispatcherFactory {
-    public Dispatcher build(Tracker tracker) {
-        return new Dispatcher(
-                new EventCache(new EventDiskCache(tracker)),
-                new Connectivity(tracker.getPiwik().getContext()),
-                new PacketFactory(tracker.getAPIUrl()
-                )
-        );
-    }
+public interface DispatcherFactory {
+    Dispatcher build(Tracker tracker);
 }
